Move database to separate module
Also move the creation of the tables into the indexer.
It's relatively cheap, but we don't need the conditional table
creation when we're trying to autocomplete values.
